add to cart, delete from cart and checkout cart

+++ objects/Carts.php
@@ -65,16 +65,16 @@ class Carts {
                 $error->code = "0001";
                 print_r(json_encode($error));
                 die();   
-                }
+            }
 
             echo "Product added to cart.";
             die();
 
-            } else {
-             $error->message = "All arguments needs a value";
-             $error->code = "0004";
-             print_r(json_encode($error));
-             die();
+        } else {
+            $error->message = "All arguments needs a value";
+            $error->code = "0004";
+            print_r(json_encode($error));
+            die();
             }
     }
 
@@ -91,8 +91,8 @@ class Carts {
         if($row = $statement->fetch()) {
              //kan användas i andra funktionen att referera till? 
             return $row['token'];
+
         } else {
-            //tänk om här kommer inte gå om det inte finns någon session för den användaren? 
             $sql = "SELECT token, TimeCreated FROM carts JOIN users ON carts.UserId = users.Id WHERE users.username = :username_IN AND TimeCreated < :active_time_IN";        
             $statement = $this->database_connection->prepare($sql);
             $statement->bindParam(":username_IN", $username_IN);
@@ -102,17 +102,83 @@ class Carts {
             $statement->bindParam(":active_time_IN", $active_time);
             $statement->execute();
 
-            $row = $statement->fetch();
+            if($row = $statement->fetch()) {
             $token = $row['token'];
 
             $sql = "DELETE FROM carts WHERE token = :token_IN";
             $statement = $this->database_connection->prepare($sql);
             $statement->bindParam(":token_IN", $token);
             $statement->execute();
+            
+            //lägg error message??
+            echo "Old session have ended and the cart have been emptied. Log in to start session";
+            die();
+            } else {
+                echo "Old session ended. Log in to start session";
+                die();
+            }
+        } 
+    }
 
-            echo "Old session ended. Log in to start session";
+    function deleteFromCart($username_IN, $product_id_IN) {
+
+        $error = new stdClass();
+        if(!empty($username_IN) && !empty($product_id_IN)) {
+        $sql = "SELECT cartitems.Id FROM cartitems JOIN carts ON cartitems.CartId = carts.Id JOIN users ON carts.UserId = users.Id WHERE username = :username_IN AND cartitems.ProductId = :product_id_IN";
+        $statement = $this->database_connection->prepare($sql);
+        $statement->bindParam(":username_IN", $username_IN);
+        $statement->bindParam(":product_id_IN", $product_id_IN);
+        $statement->execute();
+
+        $number_of_rows = $statement->rowCount();
+          
+        if($number_of_rows < 1) {
+            $error->message = "Product doesn't exist";
+            $error->code = "0008";
+            print_r(json_encode($error));
+            die();   
+        }
+        
+        $row = $statement->fetch();
+        $cart_item_id = $row['Id'];
+        
+        $sql = "DELETE FROM cartitems WHERE id = :cart_item_id_IN";
+        $statement = $this->database_connection->prepare($sql);
+        $statement->bindParam(":cart_item_id_IN", $cart_item_id);
+        $statement->execute();
+
+        $this->id = $product_id_IN;
+
+        echo "Product with id $this->id deleted.";
+        die();
+
+        } else {
+            $error->message = "All arguments needs a value";
+            $error->code = "0004";
+            print_r(json_encode($error));
             die();
         }
+    }
+
+    function checkoutCart($username_IN) {                                           
+        $sql = "SELECT * FROM carts JOIN cartitems ON carts.Id = cartitems.CartId JOIN users ON carts.UserId = users.Id JOIN products ON cartitems.ProductId = products.Id WHERE users.Username = :username_IN";
+        $statement = $this->database_connection->prepare($sql);
+        $statement->bindParam(":username_IN", $username_IN);
+        $statement->execute();
+
+        $validate_token = $this->validateToken($username_IN);
+
+        echo "<table>" . "<tr>" . "<th>Product name</th>" . "<th>Price</th>" . "<th>Quantity</th>" . "</tr>";
+        $total = 0; 
+        while($row = $statement->fetch()) {
+            $this->productname = $row['ProductName'];
+            $this->price = $row['Price'];
+            $this->quantity = $row['Quantity'];
+            $total = $total + ($row['Price'] * $row['Quantity']);
+        
+            echo "<tr>" . "<td>$this->productname</td>" . "<td>$this->price</td>" . "<td>$this->quantity</td>" . "</tr>";
+        }
+        echo "<tr>" . "<th>Total amount:</th>" .  "<td>$total</td>" . "</tr>" . "</table>";
     }
 }
 ?>
